% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{Performs a Manhattan plot.}
\usage{
manhattan.plot(chr, pos, pvalue, sig.level = NA, annotate = NULL,
  ann.default = list(), should.thin = T, thin.pos.places = 2,
  thin.logp.places = 2, xlab = "Chromosome",
  ylab = expression(-log[10](p - value)), col = c("gray", "darkgray"),
  panel.extra = NULL, pch = 20, cex = 0.8, ...)
}
\arguments{
\item{chr}{Chromosome vector of all variants}

\item{pos}{Position vector of all variants}

\item{pvalue}{P-value vector of all variants}

\item{sig.level}{significance threshold}

\item{annotate}{annotations for SNPs}

\item{ann.default}{default annotation}

\item{should.thin}{should use downsampling}

\item{thin.pos.places}{decimals used for thinning (position)}

\item{thin.logp.places}{decimals used for thinning (p-value)}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{col}{color vector}

\item{panel.extra}{extra info}

\item{pch}{plotting character}

\item{cex}{character size}

\item{...}{further arguments to xyplot, e.g. "main"}
}
\description{
Performs a Manhattan plot.
}
