% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_manhattan_and_qq_plot.R
\name{perform_manhattan_and_qq_plots}
\alias{perform_manhattan_and_qq_plots}
\title{Performs a Manhattan and QQ plot (unfiltered, MAF >1\%/INF >0.4 and MAF >5\%/INF >0.8.}
\usage{
perform_manhattan_and_qq_plots(file_pattern, out_pdf, file_sep = " ")
}
\arguments{
\item{file_pattern}{file name of results (\%CHR\% spaceholder)}

\item{out_pdf}{output PDF file name}

\item{file_sep}{field separator}
}
\description{
Performs a Manhattan and QQ plot (unfiltered, MAF >1\%/INF >0.4 and MAF >5\%/INF >0.8.
}
